added independence from `redis` and switch to `memory` storage (if `redis` is unavailable), rewritten way of forming `keyboards`, other improvements

+++ utils/database_manage/redis/clear_redis_data.py
@@ -1,11 +1,14 @@
+from aiogram.contrib.fsm_storage.redis import RedisStorage2
 from aioredis import from_url as aioredis_from_url
 
 from config import redis_data_cache
+from loader import dp
 
 
 async def clear_redis() -> None:
     """
     Deletes all keys in the redis-db.
     """
-    async with aioredis_from_url(**redis_data_cache) as connect:
-        await connect.flushall()
+    if isinstance(dp.storage, RedisStorage2):
+        async with aioredis_from_url(**redis_data_cache) as connect:
+            await connect.flushall()
